Now logging all unfulfilled expectations (including unexpected invocations).
The first unfulfilled expectation is passed to the SenTestCase in failWithException:, as usual, after logging.
This works around an issue with GTMTestCase where the first failWithException: raises an exception, terminating the running test and preventing other failures from being logged. This may or may not be an issue with vanilla SenTestCase.